+++ src/main/java/dandi/dandi/auth/config/FeignConfiguration.java
@@ -1,7 +1,7 @@
 package dandi.dandi.auth.config;
 
-import dandi.dandi.auth.adapter.out.oauth.apple.AppleApiCaller;
-import dandi.dandi.weather.adapter.out.kma.KmaWeatherApiCaller;
+import dandi.dandi.auth.adapter.out.oauth.apple.client.AppleApiCaller;
+import dandi.dandi.weather.adapter.out.kma.client.KmaWeatherApiCaller;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
